First implementation of Liquid Long Position Closure Client.
This project uses KnockoutJS and ethers as dependencies, and is built entirely with ES2015 modules and the tentative module loader import resolution stuff (shimmed in).  You must run `npm install && npm run build` in order for the project to work, but at that point you can just serve the `output` directory.